Move catalog data segment to own component
