+++ src/app/modules/it-systems/it-system-usages/it-system-usage-details/it-system-usage-details-archiving/it-system-usage-details-archiving-catalog/it-system-usage-details-archiving-catalog.component.scss
@@ -1,0 +1,16 @@
+:host {
+  display: flex;
+  flex-direction: column;
+  width: 100%;
+}
+
+.notes-column-span {
+  grid-column: span 3;
+}
+
+.catalog-link {
+  flex: 1;
+  font-size: 16px;
+  text-align: end;
+  text-decoration: underline;
+}
