+++ src/app/modules/it-systems/it-systems.module.ts
@@ -55,6 +55,7 @@ import { ItSystemHierarchyTableComponent } from './shared/it-system-hierarchy-ta
 import { ItSystemInterfacesTableComponent } from './shared/it-system-interfaces-table/it-system-interfaces-table.component';
 import { ItSystemKleOverviewComponent } from './shared/it-system-kle-overview/it-system-kle-overview.component';
 import { KleTableComponent } from './shared/kle-table/kle-table.component';
+import { ItSystemUsageDetailsArchivingCatalogComponent } from './it-system-usages/it-system-usage-details/it-system-usage-details-archiving/it-system-usage-details-archiving-catalog/it-system-usage-details-archiving-catalog.component';
 
 @NgModule({
   declarations: [
@@ -110,6 +111,7 @@ import { KleTableComponent } from './shared/kle-table/kle-table.component';
     CreateInterfaceDialogComponent,
     GdprOverviewComponent,
     GdprOverviewDialogComponent,
+    ItSystemUsageDetailsArchivingCatalogComponent,
   ],
   imports: [CommonModule, SharedModule, ITSystemsRouterModule, MatInputModule, MatProgressSpinnerModule],
 })
